Update: User policy

+++ app/Http/Controllers/UserController.php
@@ -2,56 +2,86 @@
 
 namespace App\Http\Controllers;
 
-use App\Brand;
-use App\Http\Requests\UserRequest;
 use App\User;
+use App\Brand;
+use App\Http\Requests\CreateUserRequest;
 use Illuminate\Support\Facades\Hash;
+use Symfony\Component\HttpFoundation\Response;
 
 class UserController extends Controller
 {
-    public function register(UserRequest $request)
+    /**
+     * Create new user
+     * 
+     * @param \App\Http\Requests\CreateUserRequest $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(CreateUserRequest $request)
     {
-        // TODO: check gate
-        // abort_if(Gate::denies());
-
         // Get data
-        $data = $request->all();
+        $data = $request->validated();
 
         // Hash the password
         $data['password'] = Hash::make($data['password']);
 
         // Set user role
-        $data['role'] = $request->input('role', 'BRAND_OWNER');
+        if(is_string($request->input('role')) && $request->input('role') == "super")
+            $data['is_superadmin'] = true;
+        else
+            $data['role_id'] = $request->input('role');
 
         // Create user row
         $user = User::create($data);
 
-        if($request->get('brand_id') && $data['role'] == 'BRAND_OWNER'){
-            // Attach brand 
+        // Attach brand 
+        if($request->get('brand_id') && !$user->is_superadmin){
             $brand = Brand::find($request->get('brand_id'));
             $user->brands()->attach($brand);
             $user->update(['selected_brand_id' => $brand->id]);
         }
 
         // Attach all brands to Admin
-        if($data['role'] == 'SUPER_ADMIN')
+        if($user->is_superadmin)
             $user->brands()->attach(Brand::all());
 
         return response()->success("User created successfully.", $user->load('brands'), 201);
     }
 
+    /**
+     * List users
+     * 
+     * @return \Illuminate\Http\Response
+     */
     public function index()
     {
+        abort_if(Gate::denies('list_user') || Auth()->user()->cannot('viewAny', $user), Response::HTTP_FORBIDDEN, "403 Forbidden");
+
         return response()->success("Users fetched successfully.", User::with(['brands', 'role'])->get());
     }
 
+    /**
+     * View user
+     * 
+     * @param \App\User $users
+     * @return \Illuminate\Http\Response
+     */
     public function show(User $user)
     {
+        abort_if(Gate::denies('show_user') || Auth()->user()->cannot('view', $user), Response::HTTP_FORBIDDEN, "403 Forbidden");
+
         return response()->success("User fetched successfully.", $user->load('brands'));
     }
 
-    public function destroy(User $user)
+    /**
+     * Delete user
+     * 
+     * @param \App\User $user
+     * @return \Illuminate\Http\Response
+     */
+    public function delete(User $user)
     {
+        abort_if(Gate::denies('delete_user') || Auth()->user()->cannot('delete', $user), Response::HTTP_FORBIDDEN, "403 Forbidden");
+
         // Delete user row
         $user->delete();
